+++ ScreenshotsParseTool/scripts/menu.py
@@ -2,9 +2,11 @@ import sys
 import os
 import subprocess
 import webbrowser
-from PyQt6.QtWidgets import QApplication, QMainWindow, QMessageBox, QMenuBar, QMenu
-from PyQt6.QtGui import QFontDatabase, QPixmap, QAction
-from start_parser import StartParserWindow
+import setproctitle
+from PyQt6.QtWidgets import QApplication, QMainWindow, QMessageBox
+from PyQt6.QtGui import QFontDatabase, QPixmap
+from startparser import StartParserWindow
+from ui_menu import Ui_MainWindow
 from dbase import Database
 from check_internet import check_internet_connection
 
@@ -16,34 +18,26 @@ class MainWindow(QMainWindow):
         self.path = os.path.abspath(os.path.join(self.current_dir, os.pardir))
         self.db = Database(self.path + '/database/spt_db.db')
         self.theme = self.db.check_theme()
-        if not self.theme == None:
-            if self.theme[0] == 'light':
-                from main_light import Ui_MainWindow
-                self.UI = Ui_MainWindow()
-                self.UI.setupUi(self)
-            else:
-                from main_dark import Ui_MainWindow
-                self.UI = Ui_MainWindow()
-                self.UI.setupUi(self)
+        self.UI = Ui_MainWindow()
+        self.UI.setupUi(self)
+        if self.theme == None or self.theme[0] == 'light':
+            self.setStyleSheet('QMainWindow {\n'
+            'background-color: white;\n'
+            '}')
         else:
-            from main_light import Ui_MainWindow
-            self.UI = Ui_MainWindow()
-            self.UI.setupUi(self)
+            self.setStyleSheet('QMainWindow {\n'
+            'background-color: #330230;\n'
+            '}')
         QFontDatabase.addApplicationFont(self.path + '/fonts/Rubik.ttf')
-        self.menu_bar = QMenuBar(self)
-        self.info = QMenu('SPT', self)
-        self.menu_bar.addMenu(self.info)
-        self.open_source = QAction('Open source', self)
-        self.info.addAction(self.open_source)
-        self.setMenuBar(self.menu_bar)
         self.connections()
 
     def connections(self):
-        self.UI.pushButton.clicked.connect(self.start_parser)
-        self.UI.pushButton_2.clicked.connect(self.stats)
-        self.UI.radioButton.clicked.connect(self.change_theme)
-        self.UI.pushButton_4.clicked.connect(self.terms)
-        self.open_source.triggered.connect(self.open_repository)
+        self.UI.button1.clicked.connect(self.start_parser)
+        self.UI.button2.clicked.connect(self.stats)
+        self.UI.OpenREADME.triggered.connect(self.readme)
+        self.UI.TermsofUse.triggered.connect(self.terms)
+        self.UI.ThemeLight.triggered.connect(self.change_theme)
+        self.UI.ThemeDark.triggered.connect(self.change_theme)
 
     def start_parser(self):
         if not self.db.check_terms() == None:
@@ -58,7 +52,8 @@ class MainWindow(QMainWindow):
                     self.internet_error_popup.move(400, 300)
                     self.internet_error_popup.setIcon(QMessageBox.Icon.Warning)
                     self.internet_error_popup.setText('Check your internet connection or disable VPN.')
-                    self.internet_error_popup.setDefaultButton(QMessageBox.StandardButton.Ok)
+                    self.okButton = self.internet_error_popup.addButton(QMessageBox.StandardButton.Ok)
+                    self.internet_error_popup.setDefaultButton(self.okButton)
                     self.internet_error_popup.exec()
             else:
                 self.terms()
@@ -75,7 +70,8 @@ class MainWindow(QMainWindow):
             self.stats_popup.move(400, 300)
             self.stats_popup.setIcon(QMessageBox.Icon.Information)
             self.stats_popup.setText('Screenshots parsed: ' + str(self.count) + '\nLast parse: ' + self.date)
-            self.stats_popup.setDefaultButton(QMessageBox.StandardButton.Ok)
+            self.okButton = self.stats_popup.addButton(QMessageBox.StandardButton.Ok)
+            self.stats_popup.setDefaultButton(self.okButton)
             self.stats_popup.exec()
         else:
             self.stats_popup = QMessageBox(self)
@@ -84,26 +80,37 @@ class MainWindow(QMainWindow):
             self.stats_popup.move(400, 300)
             self.stats_popup.setIcon(QMessageBox.Icon.Information)
             self.stats_popup.setText('No data yet.')
-            self.stats_popup.setDefaultButton(QMessageBox.StandardButton.Ok)
+            self.okButton = self.stats_popup.addButton(QMessageBox.StandardButton.Ok)
+            self.stats_popup.setDefaultButton(self.okButton)
             self.stats_popup.exec()
 
-    def open_repository(self):
-        webbrowser.open('https://github.com/codelao/Screenshots-Parse-Tool')
+    def readme(self):
+        webbrowser.open('https://github.com/codelao/Screenshots-Parse-Tool/blob/main/README.md')
 
     def change_theme(self):
+        sender = self.sender()
         if not self.theme == None:
             if self.theme[0] == 'light':
-                self.db.update_theme(theme='dark')
+                if not sender == self.UI.ThemeLight:
+                    self.db.update_theme(theme='dark')
+                    QApplication.exit(0)
+                    subprocess.Popen([sys.executable] + sys.argv)
+                else:
+                    pass
+            else:
+                if not sender == self.UI.ThemeDark:
+                    self.db.update_theme(theme='light')
+                    QApplication.exit(0)
+                    subprocess.Popen([sys.executable] + sys.argv)
+                else:
+                    pass
+        else:
+            if not sender == self.UI.ThemeLight:
+                self.db.add_theme(theme='dark')
                 QApplication.exit(0)
                 subprocess.Popen([sys.executable] + sys.argv)
             else:
-                self.db.update_theme(theme='light')
-                QApplication.exit(0)
-                subprocess.Popen([sys.executable] + sys.argv)
-        else:
-            self.db.add_theme(theme='dark')
-            QApplication.exit(0)
-            subprocess.Popen([sys.executable] + sys.argv)
+                self.db.add_theme(theme='light')
 
     def terms(self):
         self.terms_popup = QMessageBox(self)
@@ -112,19 +119,20 @@ class MainWindow(QMainWindow):
         self.terms_popup.move(400, 300)
         self.terms_popup.setIcon(QMessageBox.Icon.Warning)
         self.terms_popup.setText('DISCLAIMER')
-        self.terms_popup.setDetailedText('It is forbidden to use this tool for illegal or malicious purposes.\nDeveloper (Lao) is not responsible for the unethical use of this tool by other users.\n\nScreenshots Parse Tool\nv0.8.5\nLicensed under MIT')
+        self.terms_popup.setDetailedText('It is forbidden to use this tool for illegal or malicious purposes.\nDeveloper (Lao) is not responsible for the unethical use of this tool by other users.\n\nScreenshots Parse Tool\nLicensed under MIT')
         self.agreeButton = self.terms_popup.addButton("Agree", QMessageBox.ButtonRole.ActionRole)
         self.terms_popup.setDefaultButton(self.agreeButton)
         self.terms_popup.exec()
         if self.terms_popup.clickedButton() == self.agreeButton:
-            if not self.db.check_terms() == None:
-                pass
-            else:
+            if self.db.check_terms() == None:
                 self.db.add_terms(terms='agree')
                              
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
+    app.setApplicationName('Screenshots Parse Tool')
+    app.setApplicationVersion('0.11.5')
+    setproctitle.setproctitle('Screenshots Parse Tool')
     mw = MainWindow()
     mw.show()
     sys.exit(app.exec())